fix #1, fix #2 resolve #11, resolve #17, resolve #29
